Marko's BLE iOS additions made on top of 1.1-rel
The changes made in this commit are NOT tested on iOS or
Darwin. Some changes made in the 1.1_darwin branch that I deemed
uneccesary were not made and some had to be modified to fit with
content added between 1.1_darwin and 1.1-rel.

NOTE: The 1.1-rel parent branch is actually Iotivity version 1.1.1.