Changed iOS to only scan devices, not connect
This change moves the connection to updateCharacteristicsTo. After
checking that the device we wish to send a message to is in our list of
found peripherals, we try to connect to the peripheral and wait until
the peripheral is ready and send the message or, after 5 seconds, stop
trying to connect to the device and remove it from our list of found
peripherals.