Implementation of bonding for android le adapter
Main features:
* Bonding
* CCCD setting for non-default response characteristics

BONDING
-------
This commit is primarily an initial implementation of bonding for the
android le adapter. In this commit, every le connection is bonded
without actually checking if the device does in fact need encryption.
This decision was purly for implementation simplicity and will be
changed in the future.

Another thing that should change in this commit is that, as of now, each
new connection will remove, and recreate the bond instead of simply
keeping the old bond. This decision was made after noticing that my
Android 6.0.1 device was failing to recieve callbacks using an old bond.
However this may have been a false negative caused by calling discover
services too quickly after connecting. Additionaly testing is required.

CCCD Setting
------------
With the addition of multiple UUID targeting, the android le client must
also set the response characteristic's CCCD. This is done before
actually calling the writeCharacteristic method when the response
characteristics has not yet been recognized as having its CCCD set.