Send finish timeout. Remove send queue clearing.
This commit includes:
- Send finish timout
- Remove send queue clearing on GATT_ERROR
- Connection retry attempt limit of 2

Send Finish Timeout
===================
This commit includes a 20 second timout for any send request. This was
added in order to combat gatt callbacks which occasianally do not get
called. Previously, if a callback was not recieved, the client could not
continue to the next item in the send queue and was essentially dead.

Remove Send Queue Clearing
==========================
Previously, when the LE Adapter encountered a connection state change
with a GATT_ERROR (133) status, it would clear the send and receive
queues. This has been removed in this commit.

Connection Rety Attempt Limit
=============================
When the client receives a GATT_ERROR status of 133 it now only attempts
2 more connections before finishing the send.